Example - Multi-agent/agent-a/requirements.txt
a2a-sdk>=0.3.0
flask>=3.0.0
flask-cors>=4.0.0
google-genai>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0
